requirements.txt
# Graph Database
psycopg2-binary>=2.9.9

# DSPy (latest version with better compatibility)
dspy-ai>=2.5.0

# OpenAI
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0

# JSON handling
pydantic>=2.0.0
